clang-tidy: explicit narrowing conversion

+++ pmparser.c
@@ -61,10 +61,10 @@ procmaps_iterator* pmparser_parse(int pid){
 		tmp->length=(unsigned long)(tmp->addr_end-tmp->addr_start);
 		//perm
 		strcpy(tmp->perm,perm);
-		tmp->is_r=(perm[0]=='r');
-		tmp->is_w=(perm[1]=='w');
-		tmp->is_x=(perm[2]=='x');
-		tmp->is_p=(perm[3]=='p');
+		tmp->is_r=(int)(perm[0]=='r');
+		tmp->is_w=(int)(perm[1]=='w');
+		tmp->is_x=(int)(perm[2]=='x');
+		tmp->is_p=(int)(perm[3]=='p');
 
 		//offset
 		sscanf(offset,"%lx",&tmp->offset );
